double to long

+++ experiments/src/main/java/com/espertech/esper/experiments/UserDefinedFunctions.java
@@ -1,7 +1,7 @@
 package com.espertech.esper.experiments;
 
 public class UserDefinedFunctions {
-    public static double computeDolToEur(double price) {
-        return price * 0.89;
+    public static long computeDolToEur(long price) {
+        return (long) (price * 0.89);
     }
 }
